third_party/blink/web_tests/fast/events/touch/touch-target-move-documents-expected.txt

Tests a node moving documents while being touched.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Sending touchstart
Received touchstart on DIV-target
Received touchstart on #document-Main


Moving target node to new document


Sending touchmove


Re-adding target node to main document


Sending touchend
Received touchend on DIV-target
Received touchend on #document-Main


PASS successfullyParsed is true

TEST COMPLETE
